🐳 Add alternative rhel7 Dockerfile
This Dockerfile can be used to build a rhel-based image. It also
compiles the binary as part of the build, rather than that happening
externally (like in the other Dockerfile).

This is a first iteration. Future versions will likely change to a
simpler multi-stage build where the binary compile will happen in a
builder container, and get copied into the final runtime container (so
no need for COPYing the sources into the runtime container, and/or
installing (and then removing) the build dependencies).